+++ PureEdgeSim/com/mechalikh/pureedgesim/DataCentersManager/DefaultDataCenter.java
@@ -1,3 +1,23 @@
+/**
+ *     PureEdgeSim:  A Simulation Framework for Performance Evaluation of Cloud, Edge and Mist Computing Environments 
+ *
+ *     This file is part of PureEdgeSim Project.
+ *
+ *     PureEdgeSim is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     PureEdgeSim is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with PureEdgeSim. If not, see <http://www.gnu.org/licenses/>.
+ *     
+ *     @author Mechalikh
+ **/
 package com.mechalikh.pureedgesim.DataCentersManager;
 
 import java.util.List;
@@ -42,8 +62,11 @@ public class DefaultDataCenter extends DataCenter {
 
 	private void updateStatus() {
 
-		// Update energy consumption
-		updateEnergyConsumption();
+		// Check if the device is dead
+		if (getEnergyModel().isBattery()
+				&& this.getEnergyModel().getTotalEnergyConsumption() > getEnergyModel().getBatteryCapacity()) {
+			setDeath(true, simulationManager.getSimulation().clock());
+		}
 
 		// Update location
 		if (getMobilityManager().isMobile()) {
@@ -51,13 +74,5 @@ public class DefaultDataCenter extends DataCenter {
 		}
 	}
 
-	private void updateEnergyConsumption() {
-		// update the energy consumption
-		getEnergyModel().updateCpuEnergyConsumption(getResources().getCurrentCpuUtilization());
-
-		if (getEnergyModel().isBattery()
-				&& this.getEnergyModel().getTotalEnergyConsumption() > getEnergyModel().getBatteryCapacity()) {
-			setDeath(true, simulationManager.getSimulation().clock());
-		}
-	}
+ 
 }