+++ PureEdgeSim/com/mechalikh/pureedgesim/ScenarioManager/FilesParser.java
@@ -1,3 +1,23 @@
+/**
+ *     PureEdgeSim:  A Simulation Framework for Performance Evaluation of Cloud, Edge and Mist Computing Environments 
+ *
+ *     This file is part of PureEdgeSim Project.
+ *
+ *     PureEdgeSim is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     PureEdgeSim is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with PureEdgeSim. If not, see <http://www.gnu.org/licenses/>.
+ *     
+ *     @author Mechalikh
+ **/
 package com.mechalikh.pureedgesim.ScenarioManager;
 
 import java.io.File;
@@ -19,12 +39,14 @@ import com.mechalikh.pureedgesim.TasksGenerator.Application;
 public class FilesParser {
 
 	// Scan files
-	public boolean checkFiles(String simProp, String edgeDevicesFile, String edgeDataCentersFile, String appFile, String cloudFile) {
+	public boolean checkFiles(String simProp, String edgeDevicesFile, String edgeDataCentersFile, String appFile,
+			String cloudFile) {
 		SimulationParameters.EDGE_DEVICES_FILE = edgeDevicesFile;
 		SimulationParameters.EDGE_DATACENTERS_FILE = edgeDataCentersFile;
 		SimulationParameters.CLOUD_DATACENTERS_FILE = cloudFile;
 		return (checkSimulationProperties(simProp) && checkXmlFiles(edgeDevicesFile, TYPES.EDGE_DEVICE)
-				&& checkXmlFiles(edgeDataCentersFile, TYPES.EDGE_DATACENTER) && checkXmlFiles(cloudFile, TYPES.CLOUD) && checkAppFile(appFile));
+				&& checkXmlFiles(edgeDataCentersFile, TYPES.EDGE_DATACENTER) && checkXmlFiles(cloudFile, TYPES.CLOUD)
+				&& checkAppFile(appFile));
 	}
 
 	private boolean checkSimulationProperties(String simProp) {
@@ -66,7 +88,8 @@ public class FilesParser {
 					.parseBoolean(prop.getProperty("enable_orchestrators").trim());
 
 			SimulationParameters.EDGE_DEVICES_RANGE = Integer.parseInt(prop.getProperty("edge_devices_range").trim()); // meters
-			SimulationParameters.EDGE_DATACENTERS_RANGE = Integer.parseInt(prop.getProperty("edge_datacenters_coverage").trim()); // meters
+			SimulationParameters.EDGE_DATACENTERS_RANGE = Integer
+					.parseInt(prop.getProperty("edge_datacenters_coverage").trim()); // meters
 			SimulationParameters.PAUSE_LENGTH = Integer.parseInt(prop.getProperty("pause_length").trim());// seconds
 			SimulationParameters.MIN_NUM_OF_EDGE_DEVICES = Integer
 					.parseInt(prop.getProperty("min_number_of_edge_devices").trim());
@@ -79,7 +102,6 @@ public class FilesParser {
 			}
 			SimulationParameters.EDGE_DEVICE_COUNTER_STEP = Integer
 					.parseInt(prop.getProperty("edge_device_counter_size").trim());
-			SimulationParameters.SPEED = Double.parseDouble(prop.getProperty("speed").trim()); // meters per second m/s
 			SimulationParameters.BANDWIDTH_WLAN = 1000 * Integer.parseInt(prop.getProperty("wlan_bandwidth").trim()); // Mbits/s
 																														// to
 																														// Kbits/s
@@ -92,8 +114,8 @@ public class FilesParser {
 					.parseDouble(prop.getProperty("network_update_interval").trim()); // seconds
 			SimulationParameters.REALISTIC_NETWORK_MODEL = Boolean
 					.parseBoolean(prop.getProperty("realistic_network_model").trim()); // seconds
-			
-			SimulationParameters.CPU_ALLOCATION_POLICY = prop.getProperty("Applications_CPU_allocation_policy").trim(); 
+
+			SimulationParameters.CPU_ALLOCATION_POLICY = prop.getProperty("Applications_CPU_allocation_policy").trim();
 			SimulationParameters.ORCHESTRATION_ARCHITECTURES = prop.getProperty("orchestration_architectures")
 					.split(",");
 			SimulationParameters.ORCHESTRATION_AlGORITHMS = prop.getProperty("orchestration_algorithms").split(",");
@@ -141,24 +163,25 @@ public class FilesParser {
 				datacenterList = xmlDoc.getElementsByTagName("device");
 			else
 				datacenterList = xmlDoc.getElementsByTagName("datacenter");
+
 			double percentage = 0;
 			for (int i = 0; i < datacenterList.getLength(); i++) {
-
 				Node datacenterNode = datacenterList.item(i);
-
-				Element datacenterElement = (Element) datacenterNode; 
+				Element datacenterElement = (Element) datacenterNode;
 				isElementPresent(datacenterElement, "idleConsumption");
 				isElementPresent(datacenterElement, "maxConsumption");
+				
 				if (type == TYPES.EDGE_DEVICE) {
-					isElementPresent(datacenterElement, "mobility");
-					isElementPresent(datacenterElement, "battery");
-					isElementPresent(datacenterElement, "percentage");
-					percentage += Double
-							.parseDouble(datacenterElement.getElementsByTagName("percentage").item(0).getTextContent());
+					if (!checkEdgeDevice(datacenterElement))
+						return false;
+					percentage += Double.parseDouble(
+								datacenterElement.getElementsByTagName("percentage").item(0).getTextContent());
 					isElementPresent(datacenterElement, "batteryCapacity");
 					isElementPresent(datacenterElement, "generateTasks");
+					
 				} else if (type == TYPES.CLOUD) {
 					SimulationParameters.NUM_OF_CLOUD_DATACENTERS++;
+					
 				} else {
 					SimulationParameters.NUM_OF_EDGE_DATACENTERS++;
 					Element location = (Element) datacenterElement.getElementsByTagName("location").item(0);
@@ -180,14 +203,14 @@ public class FilesParser {
 					for (int k = 0; k < vmList.getLength(); k++) {
 						Node vmNode = vmList.item(k);
 
-						Element vmElement = (Element) vmNode; 
+						Element vmElement = (Element) vmNode;
 						isElementPresent(vmElement, "core");
 						isElementPresent(vmElement, "mips");
 						isElementPresent(vmElement, "ram");
 						isElementPresent(vmElement, "storage");
 					}
 				}
-			} 
+			}
 			if (percentage != 100 && type == TYPES.EDGE_DEVICE) {
 				SimLog.println(
 						"FilesParser- check the edge_devices.xml file!, the sum of percentages must be equal to 100%");
@@ -202,6 +225,45 @@ public class FilesParser {
 		return true;
 	}
 
+	private boolean checkEdgeDevice(Element datacenterElement) {
+		isElementPresent(datacenterElement, "mobility");
+		isElementPresent(datacenterElement, "battery");
+		isElementPresent(datacenterElement, "percentage");
+		isElementPresent(datacenterElement, "speed");
+		isElementPresent(datacenterElement, "minPauseDuration");
+		isElementPresent(datacenterElement, "maxPauseDuration");
+		isElementPresent(datacenterElement, "minMobilityDuration");
+		isElementPresent(datacenterElement, "maxMobilityDuration");
+		double speed = Double.parseDouble(datacenterElement.getElementsByTagName("speed").item(0).getTextContent());
+		double percentage = Double
+				.parseDouble(datacenterElement.getElementsByTagName("percentage").item(0).getTextContent());
+		if (percentage <= 0 || speed < 0 ) {
+			SimLog.println(
+					"FilesParser- check the edge_devices.xml file!, the percentage must be > 0 and the speed must be >= 0");
+			return false;
+		}
+
+		double minPauseDuration = Double
+				.parseDouble(datacenterElement.getElementsByTagName("minPauseDuration").item(0).getTextContent());
+		double maxPauseDuration = Double
+				.parseDouble(datacenterElement.getElementsByTagName("maxPauseDuration").item(0).getTextContent());
+		if (minPauseDuration > maxPauseDuration) {
+			SimLog.println(
+					"FilesParser- check the edge_devices.xml file!, the maxPauseDuration value must be greater than (or equal) the minPauseDuration value");
+			return false;
+		}
+		double minMobilityDuration = Double
+				.parseDouble(datacenterElement.getElementsByTagName("minMobilityDuration").item(0).getTextContent());
+		double maxMobilityDuration = Double
+				.parseDouble(datacenterElement.getElementsByTagName("maxMobilityDuration").item(0).getTextContent());
+		if (minMobilityDuration > maxMobilityDuration) {
+			SimLog.println(
+					"FilesParser- check the edge_devices.xml file!, the maxMobilityDuration value must be greater than (or equal) the minMobilityDuration value");
+			return false;
+		}
+		return true;
+	}
+
 	private boolean checkAppFile(String appFile) {
 		SimLog.println("FilesParser- Checking applications file");
 		Document doc;
@@ -212,7 +274,7 @@ public class FilesParser {
 			doc = dBuilder.parse(devicesFile);
 			doc.getDocumentElement().normalize();
 
-			NodeList appList = doc.getElementsByTagName("application"); 
+			NodeList appList = doc.getElementsByTagName("application");
 			for (int i = 0; i < appList.getLength(); i++) {
 				Node appNode = appList.item(i);
 
@@ -223,40 +285,40 @@ public class FilesParser {
 				isElementPresent(appElement, "request_size");
 				isElementPresent(appElement, "results_size");
 				isElementPresent(appElement, "task_length");
-				isElementPresent(appElement, "required_core"); 
-				isElementPresent(appElement, "rate"); 
-				
+				isElementPresent(appElement, "required_core");
+				isElementPresent(appElement, "rate");
+
 				// max delay in seconds
 				double max_delay = Double
 						.parseDouble(appElement.getElementsByTagName("max_delay").item(0).getTextContent());
-				
+
 				// the size of the container (KB)
 				long container_size = Long
 						.parseLong(appElement.getElementsByTagName("container_size").item(0).getTextContent());
-				
+
 				// average request size (KB)
 				long request_size = Long
 						.parseLong(appElement.getElementsByTagName("request_size").item(0).getTextContent());
-				
+
 				// average downloaded results size (KB)
 				long results_size = Long
 						.parseLong(appElement.getElementsByTagName("results_size").item(0).getTextContent());
-				
+
 				// average task length (MI)
 				double task_length = Double
 						.parseDouble(appElement.getElementsByTagName("task_length").item(0).getTextContent());
-				
+
 				// required number of CPU cores
 				int required_cores = Integer
-						.parseInt(appElement.getElementsByTagName("required_core").item(0).getTextContent()); 
-				
-				 // the size of the container (KB) 
-				int rate = Integer
-						.parseInt(appElement.getElementsByTagName("rate").item(0).getTextContent()); 
+						.parseInt(appElement.getElementsByTagName("required_core").item(0).getTextContent());
+
+				// the size of the container (KB)
+				int rate = Integer.parseInt(appElement.getElementsByTagName("rate").item(0).getTextContent());
 
 				// save apps parameters
-				SimulationParameters.APPLICATIONS_LIST.add(new Application (i,rate,max_delay,container_size,request_size,results_size,task_length,required_cores));
-			
+				SimulationParameters.APPLICATIONS_LIST.add(new Application(i, rate, max_delay, container_size,
+						request_size, results_size, task_length, required_cores));
+
 			}
 
 		} catch (Exception e) {